Remove usage of Buffer for rtree on disk serialization.
First, this PR implements a Serializer class that is used to first
compute the size of serialized data by running through the serialization
without copying to a buffer, but summing the size of all serialized
data. Then the caller can allocate the correct buffer size and
serialize the data into it using another construction of the same class
and run through the exact same serialization code.

Deserialization is done through the Deserializer class which gets
constructed on an already existing in memory buffer.

Both of these classes were used to remove the usage of buffer for the
serialization/deserialization of the rtree data to disk.

Also, the code for serializing/deserializing the delete condiiton was
simplified by using the new classes.

---
TYPE: IMPROVEMENT
DESC: Remove some usage of Buffer for on disk serialization.